requirements.txt
requests==2.31.0
pytest==7.4.0
python-dotenv==1.0.0
flask==2.0.1
werkzeug==2.0.3
openai==1.74.1
httpx>=0.24.1 